+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/accountinginvoice.go
@@ -3,7 +3,7 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/utils"
 	"github.com/ericlagergren/decimal"
 )
 
@@ -61,9 +61,20 @@ type AccountingInvoicePaymentAllocationAllocation struct {
 	// | **GBP**          | £20            | 1.277         | $25.54                     |
 	// | **EUR**          | €20            | 1.134         | $22.68                     |
 	// | **RUB**          | ₽20            | 0.015         | $0.30                      |
-	CurrencyRate *types.Decimal `json:"currencyRate,omitempty"`
+	CurrencyRate *decimal.Big `decimal:"number" json:"currencyRate,omitempty"`
 	// The total amount that has been allocated.
-	TotalAmount *types.Decimal `json:"totalAmount,omitempty"`
+	TotalAmount *decimal.Big `decimal:"number" json:"totalAmount,omitempty"`
+}
+
+func (a AccountingInvoicePaymentAllocationAllocation) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(a, "", false)
+}
+
+func (a *AccountingInvoicePaymentAllocationAllocation) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &a, "", false, false); err != nil {
+		return err
+	}
+	return nil
 }
 
 func (o *AccountingInvoicePaymentAllocationAllocation) GetAllocatedOnDate() *string {
@@ -80,14 +91,14 @@ func (o *AccountingInvoicePaymentAllocationAllocation) GetCurrency() *string {
 	return o.Currency
 }
 
-func (o *AccountingInvoicePaymentAllocationAllocation) GetCurrencyRate() *types.Decimal {
+func (o *AccountingInvoicePaymentAllocationAllocation) GetCurrencyRate() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.CurrencyRate
 }
 
-func (o *AccountingInvoicePaymentAllocationAllocation) GetTotalAmount() *types.Decimal {
+func (o *AccountingInvoicePaymentAllocationAllocation) GetTotalAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -133,13 +144,24 @@ func (o *AccountingInvoiceSalesOrderReference) GetID() *string {
 }
 
 type AccountingInvoiceWithholdingTax struct {
-	Amount types.Decimal `json:"amount"`
-	Name   string        `json:"name"`
+	Amount *decimal.Big `decimal:"number" json:"amount"`
+	Name   string       `json:"name"`
 }
 
-func (o *AccountingInvoiceWithholdingTax) GetAmount() types.Decimal {
+func (a AccountingInvoiceWithholdingTax) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(a, "", false)
+}
+
+func (a *AccountingInvoiceWithholdingTax) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &a, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *AccountingInvoiceWithholdingTax) GetAmount() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.Amount
 }
@@ -181,10 +203,10 @@ func (o *AccountingInvoiceWithholdingTax) GetName() string {
 // >
 // > In Sage 50 and ClearBooks, you may prefer to use the **invoiceNumber** to identify an invoice rather than the invoice **id**. Each time a draft invoice is submitted or printed, the draft **id** becomes void and a submitted invoice with a new **id** exists in its place. In both platforms, the **invoiceNumber** should remain the same.
 type AccountingInvoice struct {
-	AdditionalTaxAmount     *types.Decimal `json:"additionalTaxAmount,omitempty"`
-	AdditionalTaxPercentage *types.Decimal `json:"additionalTaxPercentage,omitempty"`
+	AdditionalTaxAmount     *decimal.Big `decimal:"number" json:"additionalTaxAmount,omitempty"`
+	AdditionalTaxPercentage *decimal.Big `decimal:"number" json:"additionalTaxPercentage,omitempty"`
 	// Amount outstanding on the invoice.
-	AmountDue types.Decimal `json:"amountDue"`
+	AmountDue *decimal.Big `decimal:"number" json:"amountDue"`
 	// The currency data type in Codat is the [ISO 4217](https://en.wikipedia.org/wiki/ISO_4217) currency code, e.g. _GBP_.
 	//
 	// ## Unknown currencies
@@ -218,10 +240,10 @@ type AccountingInvoice struct {
 	// | **GBP**          | £20            | 1.277         | $25.54                     |
 	// | **EUR**          | €20            | 1.134         | $22.68                     |
 	// | **RUB**          | ₽20            | 0.015         | $0.30                      |
-	CurrencyRate *types.Decimal         `json:"currencyRate,omitempty"`
+	CurrencyRate *decimal.Big           `decimal:"number" json:"currencyRate,omitempty"`
 	CustomerRef  *AccountingCustomerRef `json:"customerRef,omitempty"`
 	// Percentage rate (from 0 to 100) of discounts applied to the invoice. For example: A 5% discount will return a value of `5`, not `0.05`.
-	DiscountPercentage *types.Decimal `json:"discountPercentage,omitempty"`
+	DiscountPercentage *decimal.Big `decimal:"number" json:"discountPercentage,omitempty"`
 	// In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
 	//
 	// ```
@@ -306,37 +328,48 @@ type AccountingInvoice struct {
 	// - `Void` - An invoice can become Void when it's deleted, refunded, written off, or cancelled. A voided invoice may still be PartiallyPaid, and so all outstanding amounts on voided invoices are removed from the accounts receivable account.
 	Status InvoiceStatus `json:"status"`
 	// Total amount of the invoice excluding any taxes.
-	SubTotal *types.Decimal `json:"subTotal,omitempty"`
+	SubTotal *decimal.Big `decimal:"number" json:"subTotal,omitempty"`
 	// Supplemental data is additional data you can include in our standard data types.
 	//
 	// It is referenced as a configured dynamic key value pair that is unique to the accounting platform. [Learn more](https://docs.codat.io/using-the-api/supplemental-data/overview) about supplemental data.
 	SupplementalData *SupplementalData `json:"supplementalData,omitempty"`
 	// Amount of the invoice, inclusive of tax.
-	TotalAmount types.Decimal `json:"totalAmount"`
+	TotalAmount *decimal.Big `decimal:"number" json:"totalAmount"`
 	// Numerical value of discounts applied to the invoice.
-	TotalDiscount *types.Decimal `json:"totalDiscount,omitempty"`
+	TotalDiscount *decimal.Big `decimal:"number" json:"totalDiscount,omitempty"`
 	// Amount of tax on the invoice.
-	TotalTaxAmount types.Decimal                     `json:"totalTaxAmount"`
+	TotalTaxAmount *decimal.Big                      `decimal:"number" json:"totalTaxAmount"`
 	WithholdingTax []AccountingInvoiceWithholdingTax `json:"withholdingTax,omitempty"`
 }
 
-func (o *AccountingInvoice) GetAdditionalTaxAmount() *types.Decimal {
+func (a AccountingInvoice) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(a, "", false)
+}
+
+func (a *AccountingInvoice) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &a, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *AccountingInvoice) GetAdditionalTaxAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.AdditionalTaxAmount
 }
 
-func (o *AccountingInvoice) GetAdditionalTaxPercentage() *types.Decimal {
+func (o *AccountingInvoice) GetAdditionalTaxPercentage() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.AdditionalTaxPercentage
 }
 
-func (o *AccountingInvoice) GetAmountDue() types.Decimal {
+func (o *AccountingInvoice) GetAmountDue() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.AmountDue
 }
@@ -348,7 +381,7 @@ func (o *AccountingInvoice) GetCurrency() *string {
 	return o.Currency
 }
 
-func (o *AccountingInvoice) GetCurrencyRate() *types.Decimal {
+func (o *AccountingInvoice) GetCurrencyRate() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -362,7 +395,7 @@ func (o *AccountingInvoice) GetCustomerRef() *AccountingCustomerRef {
 	return o.CustomerRef
 }
 
-func (o *AccountingInvoice) GetDiscountPercentage() *types.Decimal {
+func (o *AccountingInvoice) GetDiscountPercentage() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -460,7 +493,7 @@ func (o *AccountingInvoice) GetStatus() InvoiceStatus {
 	return o.Status
 }
 
-func (o *AccountingInvoice) GetSubTotal() *types.Decimal {
+func (o *AccountingInvoice) GetSubTotal() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -474,23 +507,23 @@ func (o *AccountingInvoice) GetSupplementalData() *SupplementalData {
 	return o.SupplementalData
 }
 
-func (o *AccountingInvoice) GetTotalAmount() types.Decimal {
+func (o *AccountingInvoice) GetTotalAmount() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.TotalAmount
 }
 
-func (o *AccountingInvoice) GetTotalDiscount() *types.Decimal {
+func (o *AccountingInvoice) GetTotalDiscount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.TotalDiscount
 }
 
-func (o *AccountingInvoice) GetTotalTaxAmount() types.Decimal {
+func (o *AccountingInvoice) GetTotalTaxAmount() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.TotalTaxAmount
 }