+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [go v0.20.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [Go v0.20.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.20.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-09-22 00:16:59
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.89.0 (2.125.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.20.1] previous-versions/sync-for-commerce-version-1
+### Releases
+- [Go v0.20.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-commerce-version-1/v0.20.1 - previous-versions/sync-for-commerce-version-1